Database.HDBC.Query.TH

Description

This module contains templates to generate Haskell record types and HDBC instances correspond to RDB table schema.

Documentation

makeRelationalRecord

Arguments

:: Name
Type constructor name

-> Q [Dec]
Result declaration

Generate all persistable templates against defined record like type constructor.

makeRecordPersistableDefault

Arguments

:: Name
Type constructor name

-> Q [Dec]
Result declaration

Deprecated: Use makeRelationalRecord instead of this.

Deprecated. use makeRelationalRecord.

defineTableDefault'

Arguments

:: Config
Configuration to generate query with

-> String
Schema name

-> String
Table name

-> [(String, TypeQ)]
List of column name and type

-> [Name]
Derivings

-> Q [Dec]
Result declaration

Generate all HDBC templates about table except for constraint keys.

defineTableDefault

Arguments

:: Config
Configuration to generate query with

-> String
Schema name

-> String
Table name

-> [(String, TypeQ)]
List of column name and type

-> [Name]
Derivings

-> [Int]
Indexes to represent primary key

-> Maybe Int
Index of not-null key

-> Q [Dec]
Result declaration

Generate all HDBC templates about table.

defineTableFromDB'

Arguments

:: IConnection conn

=> IO conn
Connect action to system catalog database

-> Config
Configuration to generate query with

-> Driver conn
Driver definition

-> String
Schema name

-> String
Table name

-> [Name]
Derivings

-> Q [Dec]
Result declaration

Generate all HDBC templates using system catalog informations with specified config.

defineTableFromDB

Arguments

:: IConnection conn

=> IO conn
Connect action to system catalog database

-> Driver conn
Driver definition

-> String
Schema name

-> String
Table name

-> [Name]
Derivings

-> Q [Dec]
Result declaration

Generate all HDBC templates using system catalog informations.

inlineVerifiedQuery

Arguments

:: IConnection conn

=> IO conn
Connect action to system catalog database

-> Name
Top-level variable name which has Relation type

-> Relation p r
Object which has Relation type

-> Config
Configuration to generate SQL

-> QuerySuffix
suffix SQL words

-> String
Variable name to define as inlined query

-> Q [Dec]
Result declarations

Verify composed Query and inline it in compile type.