Database.HDBC.Record.Delete

Description

This module provides typed Delete running sequence which intermediate structures are typed.

Documentation

type PreparedDelete p = PreparedStatement p ()

Typed prepared delete type.

prepare :: IConnection conn => conn -> Delete p -> IO (PreparedDelete p)

Typed prepare delete operation.

prepareDelete :: IConnection conn => conn -> Delete p -> IO (PreparedDelete p)

Same as prepare.

withPrepareDelete :: IConnection conn => conn -> Delete p -> (PreparedDelete p -> IO a) -> IO a

Bracketed prepare operation.

runPreparedDelete :: ToSql SqlValue p => PreparedDelete p -> p -> IO Integer

Bind parameters, execute statement and get execution result.

runDelete :: (IConnection conn, ToSql SqlValue p) => conn -> Delete p -> p -> IO Integer

Prepare delete statement, bind parameters, execute statement and get execution result.