Database.HDBC.Record.Insert

Description

This module provides typed Insert running sequence which intermediate structures are typed.

Documentation

type PreparedInsert a = PreparedStatement a ()

Typed prepared insert type.

prepare :: IConnection conn => conn -> Insert a -> IO (PreparedInsert a)

Typed prepare insert operation.

prepareInsert :: IConnection conn => conn -> Insert a -> IO (PreparedInsert a)

Same as prepare.

runPreparedInsert :: ToSql SqlValue a => PreparedInsert a -> a -> IO Integer

Bind parameters, execute statement and get execution result.

runInsert :: (IConnection conn, ToSql SqlValue a) => conn -> Insert a -> a -> IO Integer

Prepare insert statement, bind parameters, execute statement and get execution result.

mapInsert :: (IConnection conn, ToSql SqlValue a) => conn -> Insert a -> [a] -> IO [Integer]

Prepare and insert each record.

chunksInsert :: (IConnection conn, ToSql SqlValue a) => conn -> Insert a -> [a] -> IO [[Integer]]

Prepare and insert with chunk insert statement.