Database.HDBC.Record.Query

Description

This module provides typed Query running sequence which intermediate structures are typed.

Documentation

type PreparedQuery p a = PreparedStatement p a

Typed prepared query type.

prepare

Arguments

:: IConnection conn

=> conn
Database connection

-> Query p a
Typed query

-> IO (PreparedQuery p a)
Result typed prepared query with parameter type p and result type a

Typed prepare query operation.

prepareQuery

Arguments

:: IConnection conn

=> conn
Database connection

-> Query p a
Typed query

-> IO (PreparedQuery p a)
Result typed prepared query with parameter type p and result type a

Same as prepare.

withPrepareQuery

Arguments

:: IConnection conn

=> conn
Database connection

-> Query p a
Typed query

-> (PreparedQuery p a -> IO b)
Body action to use prepared statement

-> IO b
Result action

Bracketed prepare operation.

fetch :: FromSql SqlValue a => ExecutedStatement a -> IO (Maybe a)

Fetch a record.

fetchAll :: FromSql SqlValue a => ExecutedStatement a -> IO [a]

Lazily Fetch all records.

fetchAll' :: FromSql SqlValue a => ExecutedStatement a -> IO [a]

Strict version of fetchAll.

listToUnique :: [a] -> IO (Maybe a)

Fetch expecting result records is unique.

fetchUnique :: FromSql SqlValue a => ExecutedStatement a -> IO (Maybe a)

Fetch all records but get only first record. Expecting result records is unique.

fetchUnique' :: FromSql SqlValue a => ExecutedStatement a -> IO (Maybe a)

Fetch all records but get only first record. Expecting result records is unique. Error when records count is more than one.

runStatement :: FromSql SqlValue a => BoundStatement a -> IO [a]

Execute statement and lazily fetch all records.

runStatement' :: FromSql SqlValue a => BoundStatement a -> IO [a]

Strict version of runStatement.

runPreparedQuery

Arguments

:: (ToSql SqlValue p, FromSql SqlValue a)

=> PreparedQuery p a
Statement to bind to

-> p
Parameter type

-> IO [a]
Action to get records

Bind parameters, execute statement and lazily fetch all records.

runPreparedQuery'

Arguments

:: (ToSql SqlValue p, FromSql SqlValue a)

=> PreparedQuery p a
Statement to bind to

-> p
Parameter type

-> IO [a]
Action to get records

Strict version of runPreparedQuery.

runQuery

Arguments

:: (IConnection conn, ToSql SqlValue p, FromSql SqlValue a)

=> conn
Database connection

-> Query p a
Query to get record type a requires parameter p

-> p
Parameter type

-> IO [a]
Action to get records

Prepare SQL, bind parameters, execute statement and lazily fetch all records.

runQuery'

Arguments

:: (IConnection conn, ToSql SqlValue p, FromSql SqlValue a)

=> conn
Database connection

-> Query p a
Query to get record type a requires parameter p

-> p
Parameter type

-> IO [a]
Action to get records

Strict version of runQuery.