Database.HDBC.Schema.Driver

Description

This module provides driver interface to load database system catalog via HDBC.

Documentation

type TypeMap = [(String, TypeQ)]

Mapping between type name string of DBMS and type in Haskell. Type name string depends on specification of DBMS system catalogs.

data Log

Log string type for compile time.

runLog :: (String -> t) -> (String -> t) -> Log -> t

Folding operation of Log type.

data LogChan

Channel to store compile-time warning messages.

newLogChan :: Bool -> IO LogChan

Build and return a new instance of LogChan.

takeLogs :: LogChan -> IO [Log]

Take all logs list from channel.

putWarning :: LogChan -> String -> IO ()

Push a warning string into LogChan.

putError :: LogChan -> String -> IO ()

Push an error string into LogChan.

putVerbose :: LogChan -> String -> IO ()

Put verbose compile-time message as warning when verboseAsWarning.

failWith :: LogChan -> String -> MaybeT IO a

Push an error string into LogChan and return failed context.

hoistMaybe :: Monad m => Maybe a -> MaybeT m a

Hoist from Maybe context into MaybeT.

maybeIO :: b -> (a -> b) -> MaybeT IO a -> IO b

Run MaybeT with default value.

data Driver conn

Interface type to load database system catalog via HDBC.

Constructors

Driver

Fields

typeMap :: TypeMap

Custom type mapping of this driver

getFieldsWithMap :: TypeMap -> conn -> LogChan -> String -> String -> IO ([(String, TypeQ)], [Int])

Get column name and Haskell type pairs and not-null columns index.

getPrimaryKey :: conn -> LogChan -> String -> String -> IO [String]

Get primary key column name.

emptyDriver :: IConnection conn => Driver conn

Empty definition of Driver

getFields :: IConnection conn => Driver conn -> conn -> LogChan -> String -> String -> IO ([(String, TypeQ)], [Int])

Helper function to call getFieldsWithMap using typeMap of Driver.